Fix crash when trying to recreate live url.
There is a chance that when vigoler trying to re get the urls from a
specific url youtube-dl may return partial result with error message
(video not available) with less videos then in the real list. That cause
to trying to get higher index from the result (because the previous
result was larger) and cause panic.

The fix is to use video id instead of index.