Make vigoler save the files with temporary names.
Vigoler now generate temporary file name and does not accept file name
in the api, so it is the caller responsibility to handle the name.
In cmd we rename the file after async done.
In server we leave the file with the temporary name and return the real
name, only when the client request to download the file, in the reponse.

This has two benefit:
 1. vigoler support diffrent download with the same name.
 2. Vigoler does not need to handle special character in file name.