Fix bug when video without async get deleted.
If the the server create new videos but do not start to download them,
When the video should be delete (because there is no requests),
the async is nil and server cleaner panic while printing.
Fix that problem by printing only when there is async (the printed
information related only to the async).