Fix error when there is duplicate video request.
When the user start download live video twice, the first one is the one
that running first, and because of this he will actually download the
video and create the async. The code connect between finished live video
with his parent by save pointer to the parent video and get it as parameter
when live video getting done (the live is added to the first video).
Because of the bug the second video is the one the saved in the map, so
when child live video is getting cleaned it trying to find his parent with
the id, but that parent does not contain any child ids (because is the
second video), and the clean method panic because the child does not exists
in the parent ids array (which is empty).